Add files via upload
A sample dataset has been provided with the code itself, the output segregates the tables in to list of lists. Each element in the list corresponds to the list of tables that have associations and can be joined with the other tables in the list. There is also another output that provides exact key names that associates given two tables for all tables.